Add backup slides for questions
For two potential questions that may arise, add backup slides to
explain them in a little more detail.

So as not to mess up the pacing timer calculation for the rest of the
presentation, these slides are all set to take just 1 second out of
the time budget.

Also, insert one blank slide between the main part of the presentation
and the backup slides, just so as not to advance to the question
slides accidentally.