Show speaker notes
This way, the speaker notes can be exported to a PDF, and the whole
lot printed out if necessary. This makes for a great hard-copy backup
of the deck with notes, which the speaker can use in case of A/V setup
difficulties.

For details on how to create a PDF with speaker notes, follow these
instructions:
https://revealjs.com/pdf-export/